paper-to-code-automation/requirements.txt
# Week 1 core dependencies
transformers>=4.41.0
datasets>=2.19.0
pdfplumber>=0.11.0
torch>=2.2.0
torchvision>=0.17.0
pydantic>=2.6.0
jinja2>=3.1.0
langchain>=0.2.0

# Week 2 additions (Parsing + OCR)
PyMuPDF>=1.23.8         # fitz — PDF rendering
paddleocr>=2.7.0.3      # OCR engine
tqdm>=4.66.0            # progress bars
pillow>=10.2.0          # image handling
numpy>=1.26.0           # required by paddleocr
opencv-python>=4.9.0.80 # image preprocessing for OCR

# Week 3 additions (IR Generation)
langchain-google-genai>=0.1.0  # Google's Gemini integration for LangChain
python-dotenv>=1.0.0    # for loading environment variables
litellm>=1.0.0          # Unified LLM API
google-generativeai>=0.3.0  # Google's Generative AI SDK
pytest>=7.4.0           # testing framework
pydantic>=2.0.0         # data validation and settings management

# Optional for OCR layout parsing
# layoutparser>=0.3.4    # uncomment if needed for advanced layout parsing